carousel registry done
